requirements.txt
torch
torchvision
pillow
transformers
elasticsearch
tqdm
streamlit
pandas